AbstractFactory.java, change type int to double # Please enter the commit message for your changes. Lines starting # with '#' will be kept; you may remove them yourself if you want to. # An empty message aborts the commit. # On branch master # Changes to be committed: #   (use "git reset HEAD <file>..." to unstage) # #	modified:   src/main/java/sample/design_pattern/AbstractFactory.java # # Untracked files: #   (use "git add <file>..." to include in what will be committed) # #	src/main/c/State

+++ src/main/java/sample/design_pattern/AbstractFactory.java
@@ -8,7 +8,7 @@ package sample.design_pattern;
 
 
 interface IShapeFactory {
-    IShape create(int width, int height);
+    IShape create(double width, double height);
 }
 
 interface IShape {
@@ -17,7 +17,7 @@ interface IShape {
 
 // Rectangle
 class RectangleFactory implements IShapeFactory {
-    public IShape create(int width, int height) {
+    public IShape create(double width, double height) {
         return new Rectangle(width, height);
     }
 }
@@ -38,7 +38,7 @@ class Rectangle implements IShape {
 
 // Triangle
 class TriangleFactory implements IShapeFactory {
-    public IShape create(int width, int height) {
+    public IShape create(double width, double height) {
         return new Triangle(width, height);
     }
 }
